+++ src/main/java/com/dasom/gongtalk/repository/SchoolRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SchoolRepository extends JpaRepository<School, Integer> {
+public interface SchoolRepository extends JpaRepository<School, Long> {
 
     List<School> findAll();
 }
