+++ src/main/java/com/dasom/gongtalk/repository/SettingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SettingRepository extends JpaRepository<Setting, Integer> {
+public interface SettingRepository extends JpaRepository<Setting, Long> {
 
 }
